Refactored Python parameter subclasses into pywr.parameters module.
